+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/worktime/fixedset/KshmtFixedHolRestSet.java
@@ -38,17 +38,17 @@ public class KshmtFixedHolRestSet extends UkJpaEntity implements Serializable {
     
     /** The period no. */
     @Column(name = "PERIOD_NO")
-    private short periodNo;
+    private int periodNo;
     
     
     /** The start time. */
     @Column(name = "START_TIME")
-    private short startTime;
+    private int startTime;
     
     
     /** The end time. */
     @Column(name = "END_TIME")
-    private short endTime;
+    private int endTime;
 
     /**
      * Instantiates a new kshmt fixed hol rest set.
