+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/worktime/flexset/KshmtFlexHolSetPK.java
@@ -33,7 +33,7 @@ public class KshmtFlexHolSetPK implements Serializable {
 
 	/** The worktime no. */
 	@Column(name = "WORKTIME_NO")
-	private short worktimeNo;
+	private int worktimeNo;
 
 	/**
 	 * Instantiates a new kshmt flex hol set PK.
